+++ src/util.ts
@@ -1,1 +1,3 @@
-export const createSubscriptionUrl = (feedUrl: string): string => `https://feedly.com/i/subscription/feed/${feedUrl}`;
+export function createSubscriptionUrl(feedUrl: string): string {
+  return `https://feedly.com/i/subscription/feed/${feedUrl}`;
+}